mantle/platform: move metal.go to machine/qemu/
Relocate metal.go to machine/qemu/ directory where it belongs, as it
contains QEMU-specific code for ISO installations and is only used by
ISO tests, not as a general platform implementation.
